LPS-206681 ContactBatchReindexer is an internal interface, it should not be registered as osgi service
