--- modules/apps/users-admin/users-admin-impl/src/main/java/com/liferay/users/admin/internal/search/ContactBatchReindexerImpl.java
@@ -1,66 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.users.admin.internal.search;
-
-import com.liferay.portal.kernel.dao.orm.Property;
-import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
-import com.liferay.portal.kernel.model.Contact;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.service.ClassNameLocalService;
-import com.liferay.portal.search.batch.BatchIndexingActionable;
-import com.liferay.portal.search.indexer.IndexerDocumentBuilder;
-import com.liferay.portal.search.indexer.IndexerWriter;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Luan Maoski
- */
-@Component(service = ContactBatchReindexer.class)
-public class ContactBatchReindexerImpl implements ContactBatchReindexer {
-
-	@Override
-	public void reindex(long userId, long companyId) {
-		BatchIndexingActionable batchIndexingActionable =
-			indexerWriter.getBatchIndexingActionable();
-
-		batchIndexingActionable.setAddCriteriaMethod(
-			dynamicQuery -> {
-				Property classNameIdProperty = PropertyFactoryUtil.forName(
-					"classNameId");
-
-				dynamicQuery.add(
-					classNameIdProperty.eq(
-						_classNameLocalService.getClassNameId(User.class)));
-
-				Property classPKProperty = PropertyFactoryUtil.forName(
-					"classPK");
-
-				dynamicQuery.add(classPKProperty.eq(userId));
-			});
-		batchIndexingActionable.setCompanyId(companyId);
-		batchIndexingActionable.setPerformActionMethod(
-			(Contact contact) -> batchIndexingActionable.addDocuments(
-				indexerDocumentBuilder.getDocument(contact)));
-
-		batchIndexingActionable.performActions();
-	}
-
-	@Reference(
-		target = "(indexer.class.name=com.liferay.portal.kernel.model.Contact)"
-	)
-	protected IndexerDocumentBuilder indexerDocumentBuilder;
-
-	@Reference(
-		target = "(indexer.class.name=com.liferay.portal.kernel.model.Contact)"
-	)
-	protected IndexerWriter<Contact> indexerWriter;
-
-	@Reference
-	private ClassNameLocalService _classNameLocalService;
-
-}
